Factored out common reporting steps
